fix innate item
