+++ Resources/Prototypes/Entities/Objects/Tools/handheld_mass_scanner.yml
